actualiza AuthFirebase and router file
